[Rust] Implement Experiments properly
