+++ glean-core/src/metrics/experiment.rs
@@ -2,19 +2,16 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at https://mozilla.org/MPL/2.0/.
 
-use serde::{Deserialize, Serialize};
-use serde_json::{json, Map as JsonMap, Value as JsonValue};
+use std::cmp;
 use std::collections::HashMap;
 
 use crate::error_recording::{record_error, ErrorType};
-use crate::metrics::Metric;
-use crate::metrics::MetricType;
-use crate::storage::StorageManager;
+use crate::metrics::{Metric, MetricType, RecordedExperiment};
+use crate::storage::{StorageManager, INTERNAL_STORAGE};
 use crate::util::{truncate_string_at_boundary, truncate_string_at_boundary_with_error};
 use crate::CommonMetricData;
 use crate::Glean;
 use crate::Lifetime;
-use crate::INTERNAL_STORAGE;
 
 /// The maximum length of the experiment id, the branch id, and the keys of the
 /// `extra` map. Identifiers longer than this number of characters are truncated.
@@ -26,30 +23,6 @@ const MAX_EXPERIMENT_VALUE_LEN: usize = MAX_EXPERIMENTS_IDS_LEN;
 /// beyond this limit will be dropped. Note that truncation of a hash map is
 /// nondeterministic in which items are truncated.
 const MAX_EXPERIMENTS_EXTRAS_SIZE: usize = 20;
-
-/// The data for a single experiment.
-#[derive(Debug, Clone, Deserialize, Serialize, PartialEq, Eq)]
-pub struct RecordedExperimentData {
-    pub branch: String,
-    pub extra: Option<HashMap<String, String>>,
-}
-
-impl RecordedExperimentData {
-    /// Gets the recorded experiment data as a JSON value.
-    ///
-    /// For JSON, we don't want to include `{"extra": null}` -- we just want to skip
-    /// `extra` entirely. Unfortunately, we can't use a serde field annotation for this,
-    /// since that would break bincode serialization, which doesn't support skipping
-    /// fields. Therefore, we use a custom serialization function just for JSON here.
-    pub fn as_json(&self) -> JsonValue {
-        let mut value = JsonMap::new();
-        value.insert("branch".to_string(), json!(self.branch));
-        if self.extra.is_some() {
-            value.insert("extra".to_string(), json!(self.extra));
-        }
-        JsonValue::Object(value)
-    }
-}
 
 /// An experiment metric.
 ///
@@ -73,6 +46,10 @@ impl ExperimentMetric {
     ///
     /// * `id` - the id of the experiment. Please note that this will be
     ///          truncated to `MAX_EXPERIMENTS_IDS_LEN`, if needed.
+    ///
+    /// # Implementation note
+    ///
+    /// This runs synchronously and queries the database to record potential errors.
     pub fn new(glean: &Glean, id: String) -> Self {
         let mut error = None;
 
@@ -123,12 +100,7 @@ impl ExperimentMetric {
     ///               truncated to `MAX_EXPERIMENTS_IDS_LEN`, if needed.
     /// * `extra` - an optional, user defined String to String map used to provide richer
     ///             experiment context if needed.
-    pub fn set_active(
-        &self,
-        glean: &Glean,
-        branch: String,
-        extra: Option<HashMap<String, String>>,
-    ) {
+    pub fn set_active_sync(&self, glean: &Glean, branch: String, extra: HashMap<String, String>) {
         if !self.should_record(glean) {
             return;
         }
@@ -146,47 +118,45 @@ impl ExperimentMetric {
         };
 
         // Apply limits to extras
-        let truncated_extras = extra.map(|extra| {
-            if extra.len() > MAX_EXPERIMENTS_EXTRAS_SIZE {
-                let msg = format!(
-                    "Extra hash map length {} exceeds maximum of {}",
-                    extra.len(),
-                    MAX_EXPERIMENTS_EXTRAS_SIZE
-                );
-                record_error(glean, &self.meta, ErrorType::InvalidValue, msg, None);
-            }
+        if extra.len() > MAX_EXPERIMENTS_EXTRAS_SIZE {
+            let msg = format!(
+                "Extra hash map length {} exceeds maximum of {}",
+                extra.len(),
+                MAX_EXPERIMENTS_EXTRAS_SIZE
+            );
+            record_error(glean, &self.meta, ErrorType::InvalidValue, msg, None);
+        }
 
-            let mut temp_map = HashMap::new();
-            for (key, value) in extra.into_iter().take(MAX_EXPERIMENTS_EXTRAS_SIZE) {
-                let truncated_key = if key.len() > MAX_EXPERIMENTS_IDS_LEN {
-                    truncate_string_at_boundary_with_error(
-                        glean,
-                        &self.meta,
-                        key,
-                        MAX_EXPERIMENTS_IDS_LEN,
-                    )
-                } else {
-                    key
-                };
-                let truncated_value = if value.len() > MAX_EXPERIMENT_VALUE_LEN {
-                    truncate_string_at_boundary_with_error(
-                        glean,
-                        &self.meta,
-                        value,
-                        MAX_EXPERIMENT_VALUE_LEN,
-                    )
-                } else {
-                    value
-                };
+        let mut truncated_extras =
+            HashMap::with_capacity(cmp::min(extra.len(), MAX_EXPERIMENTS_EXTRAS_SIZE));
+        for (key, value) in extra.into_iter().take(MAX_EXPERIMENTS_EXTRAS_SIZE) {
+            let truncated_key = if key.len() > MAX_EXPERIMENTS_IDS_LEN {
+                truncate_string_at_boundary_with_error(
+                    glean,
+                    &self.meta,
+                    key,
+                    MAX_EXPERIMENTS_IDS_LEN,
+                )
+            } else {
+                key
+            };
+            let truncated_value = if value.len() > MAX_EXPERIMENT_VALUE_LEN {
+                truncate_string_at_boundary_with_error(
+                    glean,
+                    &self.meta,
+                    value,
+                    MAX_EXPERIMENT_VALUE_LEN,
+                )
+            } else {
+                value
+            };
 
-                temp_map.insert(truncated_key, truncated_value);
-            }
-            temp_map
-        });
+            truncated_extras.insert(truncated_key, truncated_value);
+        }
 
-        let value = Metric::Experiment(RecordedExperimentData {
+        let value = Metric::Experiment(RecordedExperiment {
             branch: truncated_branch,
-            extra: truncated_extras,
+            extra: Some(truncated_extras),
         });
         glean.storage().record(glean, &self.meta, &value)
     }
@@ -196,7 +166,7 @@ impl ExperimentMetric {
     /// # Arguments
     ///
     /// * `glean` - The Glean instance this metric belongs to.
-    pub fn set_inactive(&self, glean: &Glean) {
+    pub fn set_inactive_sync(&self, glean: &Glean) {
         if !self.should_record(glean) {
             return;
         }
@@ -213,17 +183,17 @@ impl ExperimentMetric {
     /// **Test-only API (exported for FFI purposes).**
     ///
     /// Gets the currently stored experiment data as a JSON representation of
-    /// the RecordedExperimentData.
+    /// the RecordedExperiment.
     ///
     /// This doesn't clear the stored value.
-    pub fn test_get_value_as_json_string(&self, glean: &Glean) -> Option<String> {
+    pub fn test_get_value(&self, glean: &Glean) -> Option<RecordedExperiment> {
         match StorageManager.snapshot_metric_for_test(
             glean.storage(),
             INTERNAL_STORAGE,
             &self.meta.identifier(glean),
             self.meta.lifetime,
         ) {
-            Some(Metric::Experiment(e)) => Some(json!(e).to_string()),
+            Some(Metric::Experiment(e)) => Some(e),
             _ => None,
         }
     }
@@ -235,14 +205,14 @@ mod test {
 
     #[test]
     fn stable_serialization() {
-        let experiment_empty = RecordedExperimentData {
+        let experiment_empty = RecordedExperiment {
             branch: "branch".into(),
-            extra: None,
+            extra: Default::default(),
         };
 
         let mut data = HashMap::new();
         data.insert("a key".to_string(), "a value".to_string());
-        let experiment_data = RecordedExperimentData {
+        let experiment_data = RecordedExperiment {
             branch: "branch".into(),
             extra: Some(data),
         };
@@ -264,9 +234,9 @@ mod test {
     #[rustfmt::skip] // Let's not add newlines unnecessary
     fn deserialize_old_encoding() {
         // generated by `bincode::serialize` as of Glean commit ac27fceb7c0d5a7288d7d569e8c5c5399a53afb2
-        // empty was generated from: `RecordedExperimentData { branch: "branch".into(), extra: None, }`
+        // empty was generated from: `RecordedExperiment { branch: "branch".into(), extra: None, }`
         let empty_bin = vec![6, 0, 0, 0, 0, 0, 0, 0, 98, 114, 97, 110, 99, 104];
-        // data was generated from: RecordedExperimentData { branch: "branch".into(), extra: Some({"a key": "a value"}), };
+        // data was generated from: RecordedExperiment { branch: "branch".into(), extra: Some({"a key": "a value"}), };
         let data_bin  = vec![6, 0, 0, 0, 0, 0, 0, 0, 98, 114, 97, 110, 99, 104,
                              1, 1, 0, 0, 0, 0, 0, 0, 0, 5, 0, 0, 0, 0, 0, 0, 0,
                              97, 32, 107, 101, 121, 7, 0, 0, 0, 0, 0, 0, 0, 97,
@@ -275,11 +245,11 @@ mod test {
 
         let mut data = HashMap::new();
         data.insert("a key".to_string(), "a value".to_string());
-        let experiment_data = RecordedExperimentData { branch: "branch".into(), extra: Some(data), };
+        let experiment_data = RecordedExperiment { branch: "branch".into(), extra: Some(data), };
 
         // We can't actually decode old experiment data.
         // Luckily Glean did store experiments in the database before commit ac27fceb7c0d5a7288d7d569e8c5c5399a53afb2.
-        let experiment_empty: Result<RecordedExperimentData, _> = bincode::deserialize(&empty_bin);
+        let experiment_empty: Result<RecordedExperiment, _> = bincode::deserialize(&empty_bin);
         assert!(experiment_empty.is_err());
 
         assert_eq!(experiment_data, bincode::deserialize(&data_bin).unwrap());
